Add files via upload
Cluster-Specific Gene Expression: Genes that are highly expressed in one or a few clusters (e.g., brighter colors in one column) may be key markers of the unique identity and function of those cell populations are visualized in this plot. 

In the context to the Functional Annotation,  Clusters with similar expression patterns for a group of genes such as similar color patterns across rows,  might be functionally related or share similar pathways or physiological states.


In the content of the Gene Co-expression, this plots also indicates co-expressed genes, which may be regulated together or participate in common biological processes or pathways.

Ultimately, the heterogeneity of cell populations in a sample and for inferring the potential biological significance of gene expression patterns within the context of single-cell RNA sequencing.